requirements.txt
SQLAlchemy==1.3.18
mysql-connector==2.2.9
python-dotenv==0.13.0